Unify SynqOrderService and SynqProductService
